services/api/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "AURA365 Backend API - Django REST Framework"
readme = "README.md"
requires-python = ">=3.13.7"
dependencies = [
    "django>=5.2.7",
    "django-filter>=25.1",
    "djangorestframework>=3.16.1",
    "markdown>=3.9",
    "psycopg2-binary>=2.9.10",
    "python-decouple>=3.8",
    "django-cors-headers>=4.6.0",
    "drf-spectacular>=0.28.0",
    "cryptography>=43.0.0",
    "pyjwt[crypto]>=2.9",
    "requests>=2.32.3",
    "responses>=0.25.3",
    "celery>=5.5.3",
    "redis>=7.0.1",
    "gunicorn>=23.0.0",
    "aura360-messaging @ file:///Users/freakscode/Proyectos%202025/AURA360/services/shared",
    "qdrant-client>=1.11.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.4.2",
    "pytest-django>=4.9.0",
    "pytest-cov>=6.0.0",
]

[build-system]
requires = ["setuptools>=65.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["config*", "users*", "holistic*", "body*"]
